+++ includes/execution.h
@@ -6,7 +6,7 @@
 /*   By: hnakayam <hnakayam@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/09 14:26:05 by retanaka          #+#    #+#             */
-/*   Updated: 2024/11/24 18:08:11 by hnakayam         ###   ########.fr       */
+/*   Updated: 2024/11/24 19:13:54 by hnakayam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,13 +47,43 @@ char		*search_cmd(char *cmd_without_op, char **splited_path_envp);
 char		*get_path_cmd(char *cmd, char **splited_path_envp);
 void		free_two_dimensional_array(char **strs);
 
-// excute
-void		excute_cmd(t_cmd *cmd, char **splited_path_env,
-				t_env **env, int status);
+// execute_cmds
+int			is_builtin(char *cmd);
+int			destruct_forks(t_cmd *cmds, size_t len);
+t_cmd		*init_cmds(t_cmd_a *cmd_a_s);
 int			exe_cmds(t_cmd_a *cmd_a_s, t_env **env, int *status);
+
+// execute_cmds
 int			execute_builtin_cmd(t_env **env, t_cmd *cmd, int status,
 				int is_child);
-int			is_builtin(char *cmd);
+void		excute_cmd(t_cmd *cmd, char **splited_path_env,
+				t_env **env, int status);
+void		check_is_file(char *path_cmd, char *cmd);
+
+// expand
+void		delete_cmd_exe(t_cmd *cmd);
+size_t		count_tkns_for_word(t_tkn *tkns);
+void		count_cmd(size_t *words_len, size_t *reds_len, t_tkn *tkns);
+int			ft_is_charactor(char c);
+size_t		charactors_len(const char *src);
+size_t		check_valiable(const char *src);
+char		*get_value(const char *src, t_env *env, char *st);
+void		count_word_in_a_tkn(t_tkn *tkn, size_t *word_len, t_env *env,
+				char *st);
+size_t		count_word(t_tkn *tkns, t_env *env, char *st);
+void		store_word_in_a_tkn(char *dst, t_tkn *tkn, t_env *env, char *st);
+void		store_word(char *dst, t_tkn *tkns, t_env *env, char *st);
+int			create_word(char **pp, t_tkn *tkns, t_env *env, char *st);
+int			is_env_variable(const char *head, const char *tail);
+size_t		count_heredoc(t_tkn *tkns, size_t t_len);
+size_t		store_heredoc(t_red *red, t_tkn *tkns, size_t t_len);
+int			create_heredoc(t_red *red, t_tkn *tkns);
+int			store_cmd(t_cmd *cmd, t_tkn *tkns, t_env *env, char *st);
+t_cmd		*expand_cmd(t_cmd *cmd, t_tkn *tkns, t_env *env, int status);
+
+// execute_utils
+size_t		count_env_len(t_env *env);
+char		**env_to_environ(t_env *env);
 
 // pipe
 int			prepare_pipe(t_cmd *cmd);
