+++ srcs/execution/excute.c
@@ -6,7 +6,7 @@
 /*   By: hnakayam <hnakayam@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/13 14:17:03 by hnakayam          #+#    #+#             */
-/*   Updated: 2024/11/24 18:01:45 by hnakayam         ###   ########.fr       */
+/*   Updated: 2024/11/24 19:12:00 by hnakayam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -408,257 +408,4 @@ t_cmd	*expand_cmd(t_cmd *cmd, t_tkn *tkns, t_env *env, int status)
 	if (!store_cmd(cmd, tkns, env, st))
 		return (free(st), NULL);
 	return (free(st), cmd);
-}
-
-size_t	count_env_len(t_env *env)
-{
-	size_t	len;
-
-	len = 0;
-	while (env != NULL)
-	{
-		env = env->next;
-		len++;
-	}
-	return (len);
-}
-
-char	**env_to_environ(t_env *env)
-{
-	char	**strs;
-	char	*tmp;
-	size_t	i;
-
-	strs = (char **)ft_xcalloc(sizeof(char *) * (count_env_len(env) + 1));
-	if (strs == NULL)
-		return (NULL);
-	i = 0;
-	while (env != NULL && env->value != NULL)
-	{
-		strs[i] = ft_strjoin(env->key, "=");
-		if (strs[i] == NULL)
-			return (free_two_dimensional_array(strs), NULL);
-		tmp = strs[i];
-		strs[i] = ft_strjoin(strs[i], env->value);
-		if (strs[i] == NULL)
-			return (free_two_dimensional_array(strs), NULL);
-		free(tmp);
-		env = env->next;
-		i++;
-	}
-	strs[i] = NULL;
-	return (strs);
-}
-
-void	check_is_file(char *path_cmd, char *cmd)
-{
-	struct stat		st;
-
-	if (strcmp(cmd, "") == 0)
-	{
-		ft_printf_stderr("Command '' not found\n");
-		exit(127);
-	}
-	if (stat(path_cmd, &st) < 0)
-	{
-		perror("stat");
-		exit(1);
-	}
-	if (st.st_mode & S_IFDIR && !(st.st_mode & S_IFREG))
-	{
-		ft_printf_stderr("minishell: %s: Is a directory\n", cmd);
-		exit(126);
-	}
-}
-
-void	excute_cmd(t_cmd *cmd, char **splited_path_env, t_env **env, int status)
-{
-	char	*path_cmd;
-	char	**environ;
-	char	*st;
-
-	set_exec_child_handler();
-	prepare_pipe_in_child(cmd);
-
-	st = ft_itoa(status);
-	if (!st)
-		exit(1);
-	if (!open_all_file(cmd->reds, 1, *env, st))
-		exit(1);
-	free(st);
-	if (cmd->reds != NULL)
-		set_redirects(cmd->reds);
-	if (cmd->words[0] == NULL) // redirectのみのときは終了
-		exit(0);
-	path_cmd = get_path_cmd(cmd->words[0], splited_path_env);
-	// check_is_file
-	check_is_file(path_cmd, cmd->words[0]);
-	environ = env_to_environ(*env);
-	if (environ == NULL)
-		exit(1);
-	execve(path_cmd, cmd->words, environ);
-	ft_printf_stderr("Error : execve\n");
-	exit(1);
-}
-
-int	execute_builtin_cmd(t_env **env, t_cmd *cmd, int status, int is_child)
-{
-	char	*command_name;
-	char	**args;
-	char	*st;
-
-	if (is_child)
-	{
-		prepare_pipe_in_child(cmd);
-	}
-	st = ft_itoa(status);
-	if (!st)
-		return (1);
-	if (!open_all_file(cmd->reds, is_child, *env, st))
-		return (1); // 1 ではない気がする
-	free(st);
-	if (cmd->reds != NULL)
-		set_redirects(cmd->reds);
-	command_name = cmd->words[0];
-	args = &cmd->words[1];
-	if (strcmp(command_name, "cd") == 0)
-		return (builtin_cd(env, args));
-	else if (strcmp(command_name, "pwd") == 0)
-		return (builtin_pwd());
-	else if (strcmp(command_name, "echo") == 0)
-		return (builtin_echo(args));
-	else if (strcmp(command_name, "env") == 0)
-		return (builtin_env(env, args));
-	else if (strcmp(command_name, "exit") == 0)
-		return (builtin_exit(*env, args, status, is_child));
-	else if (strcmp(command_name, "export") == 0)
-		return (builtin_export(env, args));
-	else if (strcmp(command_name, "unset") == 0)
-		return (builtin_unset(env, args));
-	else // fatal_error
-		return (1);
-}
-
-int	is_builtin(char *cmd)
-{
-	if (cmd == NULL)
-		return (0);
-	if (strcmp(cmd, "cd") == 0 || strcmp(cmd, "echo") == 0 || \
-		strcmp(cmd, "env") == 0 || strcmp(cmd, "exit") == 0 || \
-		strcmp(cmd, "export") == 0 || strcmp(cmd, "pwd") == 0 || \
-		strcmp(cmd, "unset") == 0)
-	{
-		return (1);
-	}
-	return (0);
-}
-
-int	destruct_forks(t_cmd *cmds, size_t len)
-{
-	size_t	i;
-	int		status;
-	int		return_status;
-
-	i = 0;
-	status = 0;
-	return_status = 0;
-	while (i < len)
-	{
-		waitpid(cmds[i].pid, &status, 0);
-		if (WIFSIGNALED(status))
-		{
-			return_status = WTERMSIG(status) + 128;
-			if (return_status == 130)
-				ft_printf_stderr("\n");
-		}
-		if (WIFEXITED(status))
-			return_status = WEXITSTATUS(status);
-		i++;
-	}
-	// free(cmds);
-	return (return_status);
-}
-
-t_cmd	*init_cmds(t_cmd_a *cmd_a_s)
-{
-	t_cmd	*cmds;
-	size_t	len;
-
-	len = 0;
-	while (cmd_a_s[len].tkns)
-		len++;
-	cmds = ft_xcalloc(sizeof(t_cmd) * (len + 1));
-	if (cmds == NULL)
-		return (NULL);
-	cmds[len].type = TAIL;
-	while (len--)
-	{
-		cmds[len].type = NORMAL;
-		cmds[len].pipe_in[0] = 0;
-		cmds[len].pipe_in[1] = -1;
-		cmds[len].pipe_out[0] = -1;
-		cmds[len].pipe_out[1] = 1;
-		cmds[len].pid = 0;
-	}
-	return (cmds);
-}
-
-int	exe_cmds(t_cmd_a *cmd_a_s, t_env **env, int *status)
-{
-	size_t	i;
-	int		tmp_in; // test
-	int		tmp_out; // test
-	char	**splited_path_env;
-	t_cmd	*cmds;
-
-	splited_path_env = get_splited_path_env(*env);
-	cmds = init_cmds(cmd_a_s);
-	if (cmds == NULL)
-		return (1); // free
-	i = 0;
-	while (cmds[i].type != TAIL)
-	{
-		tmp_in = dup(0); // test
-		tmp_out = dup(1); // test
-		if (prepare_pipe(&cmds[i]))
-			break ;
-		if (!expand_cmd(&cmds[i], cmd_a_s[i].tkns, *env, *status))
-			return (0);
-		if (is_builtin(cmds[i].words[0]) && i == 0 && cmds[i + 1].type == TAIL)
-		{
-			*status = execute_builtin_cmd(env, &cmds[i], *status, 0);
-			free_two_dimensional_array(splited_path_env);
-			delete_cmd_exe(cmds);
-			free(cmds);
-			dup2(tmp_in, 0);
-			dup2(tmp_out, 1);
-			close(tmp_in);
-			close(tmp_out);
-			return (*status); // unnecessary ?
-		}
-		cmds[i].pid = fork();
-		if (cmds[i].pid < 0)
-		{
-			ft_printf_stderr("Error: %s: %s\n", "fork", strerror(errno));
-			break ;
-		}
-		else if (cmds[i].pid == 0 && is_builtin(cmds[i].words[0]))
-			exit(execute_builtin_cmd(env, &cmds[i], *status, 1));
-		else if (cmds[i].pid == 0)
-		{
-			excute_cmd(&cmds[i], splited_path_env, env, *status);
-		}
-		set_exec_handler(true);
-		prepare_pipe_in_parent(&cmds[i]);
-		dup2(tmp_in, 0);
-		dup2(tmp_out, 1);
-		close(tmp_in);
-		close(tmp_out);
-		delete_cmd_exe(&cmds[i]);
-		i++;
-	}
-	*status = destruct_forks(cmds, i);
-	free_two_dimensional_array(splited_path_env);
-	free(cmds);
-	return (*status); // unnecessary ?
 }
